requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.1
fire==0.6.0
fonttools==4.53.0
fsspec==2024.2.0
idna==3.7
importlib_resources==6.4.0
intel-openmp==2021.4.0
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.2.0
pip==24.1.1
psutil==5.9.8
py-cpuinfo==9.0.0
pybboxes==0.1.6
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
sahi==0.10.0
scikit-learn==1.2.2
scipy==1.13.1
seaborn==0.13.2
setuptools==58.1.0
shapely==2.0.4
six==1.16.0
sympy==1.12
tbb==2021.11.0
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
torch==2.3.1+cu118
torchaudio==2.3.1+cu118
torchvision==0.18.1+cu118
tqdm==4.66.4
typing_extensions==4.9.0
tzdata==2024.1
ultralytics==8.2.28
ultralytics-thop==0.2.7
urllib3==2.2.1
zipp==3.19.2
